refactor: allocations eventual, allocation signer, sync interval
