+++ packages/indexer-service/src/commands/start.ts
@@ -22,6 +22,7 @@ import {
   indexerError,
   IndexerErrorCode,
   GraphNode,
+  monitorEligibleAllocations,
   Network,
   NetworkSubgraph,
   registerIndexerErrorMetrics,
@@ -31,7 +32,7 @@ import {
 
 import { createServer } from '../server'
 import { QueryProcessor } from '../queries'
-import { ensureAttestationSigners, monitorEligibleAllocations } from '../allocations'
+import { ensureAttestationSigners } from '../allocations'
 import { AllocationReceiptManager } from '../query-fees'
 import pRetry from 'p-retry'
